Fix locking mechanism for _replace_files
The update method will download files and attempt to replace them if the
code is not running. If it is running, the _replace_files method now
returns false. This also creates a file called ".queue" to signal whether
file replacement is still necessary.

TODO: Make sure version number is not updated too early